Make CurrentPathwayProvider update PathwayProvider
Editing the pathway used to change the PathwayProvider but now it just
changes the current pathway which in turn updates the pathway provider.
Undo appears to be working now.
Only difference in behavior is that when the nodeid in the url does not
exist it will default to start but will not update the url